Re #17839 Moved new code out of routines into new file
