+++ scripts/CMakeLists.txt
@@ -35,6 +35,8 @@ set ( TEST_PY_FILES
       test/ISISDirecInelasticConfigTest.py
       test/IndirectCommonTests.py
       test/PearlRoutinesTest.py
+      test/PearlPowder_commonTest.py
+      test/PearlPowder_PEARLTest.py
       test/PyChopTest.py
       test/ReductionWrapperTest.py
       test/ReflectometryQuickAuxiliaryTest.py
